add-admin-region-editing: added the scoping to the top to allow to allow merging of queries. Just a small him and haw now with that.
added the scoping to the top to allow to allow merging of queries.
Just a small him and haw now with that do I add some kind of rule to allow only one query of every type lol? Really its pointless to add more than one query of any given type to a query.
The scopes already throw with this due to the scoping but then do I add some kind of rule to throw out an explanation as well?